+++ templates/react-app/nwb.config.js
@@ -1,4 +1,3 @@
 module.exports = {
-  // Let nwb know this is a React app when generic build commands are used
   type: 'react-app'
 }