+++ tests/fixtures/0.8-build-config-compat.js
@@ -1,0 +1,7 @@
+module.exports = {
+  type: 'react-component',
+  externals: {react: 'React'},
+  global: 'Test',
+  jsNext: true,
+  umd: true
+}