+++ tests/fixtures/partial-build-config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  type: 'web-module',
+  build: {
+    umd: true,
+    global: 'Test'
+  }
+}